AYS-156 | `AysConfigurationException` Class Has Been Moved to `AysConfigurationProperty` Class
